1.Добавил удаление позиции из корзины. 2.Написал аспекты для логирования всех методов. 3.Написал аннотацию @Timer для использования в методах для измерения времени выполнения этого метода.

+++ src/main/java/ru/geekbrains/spring2/controllers/CartController.java
@@ -28,7 +28,7 @@ public class CartController {
     }
 
     @GetMapping("/delete/{id}")
-    public void deleteProductsById(@PathVariable int id) {
+    public void deleteProductsById(@PathVariable Long id) {
         cartService.deleteById(id);
     }
 }
